swtcon: Add host-buildable unit tests for the ABI-width fixes
Widens WorkItem::pixelDataPtr and the state-buffer path pointer to
uintptr_t so they round-trip losslessly on a 64-bit dev host (the prior
int32_t truncated real heap addresses, corrupting commit_item's pixel
read). Gates the still-32-bit-only library-impl dlopen path behind
SWTCON_32BIT_ABI_CHECK so it compiles out entirely on non-armv7 hosts
instead of just being runtime-dead.

Adds MockFb, a linker-wrapped fake /dev/fb0 + ioctl responder, letting
init_framebuffer/pan_and_unblank/blank_fb's real success and failure
branches run end-to-end on the dev host, and a large expansion of
TestSwtcon.cpp covering swtcon_init/shutdown, the update queue/batching
pipeline, commit_item/dispatch_processed_regions, and the worker/display
threads driven against the mock. Also exposes pan_and_advance_frame for
direct testing and strengthens several existing assertions (subtract_
update_region's split strips, commit_item's packed transitions, save_
statebuffer's/write_lut_pattern's/fill_var_screeninfo's actual output
content) that previously only checked sizes or non-null pointers rather
than the values themselves.

+++ libs/swtcon/include/swtcon.h
@@ -105,8 +105,11 @@ void swtcon_wait();
 void swtcon_suspend();
 void swtcon_resume();
 
-// Re-implemented natively
-void swtcon_shutdown(int state_ptr_or_zero);
+// Re-implemented natively. state_ptr_or_zero is 0, or a raw pointer-sized
+// int naming the file to save the statebuffer to (see save_statebuffer's
+// own comment in init.h) - uintptr_t rather than int32_t so the round-trip
+// stays lossless on a 64-bit host too.
+void swtcon_shutdown(uintptr_t state_ptr_or_zero);
 
 // Runtime load bias of libqsgepaper.so: ghidra_addr = runtime_pc - offset.
 uintptr_t swtcon_runtime_offset();
